Moved classes into Utility module.
