--- lib/concurrent/utilities.rb
@@ -1,4 +1,0 @@
-require 'concurrent/utility/monotonic_time'
-require 'concurrent/utility/processor_count'
-require 'concurrent/utility/timeout'
-require 'concurrent/utility/timer'
