+++ lib/concurrent/utility/timer.rb
@@ -1,13 +1,9 @@
 require 'concurrent/configuration'
-require 'thread'
 
 module Concurrent
 
-  # Perform the given operation asynchronously after the given number of seconds.
-  #
-  # This is a convenience method for posting tasks to the global timer set.
-  # It is intended to be simple and easy to use. For greater control use
-  # either `TimerSet` or `ScheduledTask` directly.
+  # [DEPRECATED] Perform the given operation asynchronously after
+  # the given number of seconds.
   #
   # @param [Fixnum] seconds the interval in seconds to wait before executing the task
   #
@@ -16,10 +12,10 @@ module Concurrent
   # @return [Concurrent::ScheduledTask] IVar representing the task
   #
   # @see Concurrent::ScheduledTask
-  # @see Concurrent::TimerSet
   #
-  # @!macro monotonic_clock_warning
+  # @deprecated use `ScheduledTask` instead
   def timer(seconds, *args, &block)
+    warn '[DEPRECATED] use ScheduledTask instead'
     raise ArgumentError.new('no block given') unless block_given?
     raise ArgumentError.new('interval must be greater than or equal to zero') if seconds < 0
     Concurrent.configuration.global_timer_set.post(seconds, *args, &block)
